requirements.txt
openai>=1.12.0
python-dotenv>=1.0.0
sounddevice>=0.4.6
numpy>=1.24.0
scipy>=1.11.0
soundfile>=0.12.1 